Feat: [EVER-103] 쿠폰, 브랜드 엔티티 수정 및 구조 변경

+++ src/main/java/com/team4ever/backend/domain/coupon/dto/CouponUseResponse.java
@@ -3,12 +3,11 @@ package com.team4ever.backend.domain.coupon.dto;
 import com.team4ever.backend.domain.coupon.entity.UserCoupon;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
 public class CouponUseResponse {
-    private Long couponId;
+    private Integer couponId;
     private Boolean isUsed;
 
     public static CouponUseResponse from(UserCoupon uc) {
